+++ CRM-master/C R M/Controllers/EmpresasController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Data;
 using System.Data.Entity;
+using System.Data.SqlClient;
 using System.Linq;
 using System.Net;
 using System.Web;
@@ -10,6 +11,7 @@ using C_R_M.Models;
 
 namespace C_R_M.Controllers
 {
+    [PermisoAttribute]
     public class EmpresasController : Controller
     {
         private CRMEntities db = new CRMEntities();
@@ -17,13 +19,18 @@ namespace C_R_M.Controllers
         // GET: Empresas
         public ActionResult Index()
         {
-            var empresa = db.Empresa.Include(e => e.Canton).Include(e => e.Distrito).Include(e => e.Provincia).Include(e => e.Pais1.Nombre);
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
+            var empresa = db.Empresa.Include(e => e.Canton).Include(e => e.Distrito).Include(e => e.Provincia).Include(e => e.Pais);
             return View(empresa.ToList());
         }
+
 
         // GET: Empresas/Details/5
         public ActionResult Details(int? id)
         {
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -39,11 +46,14 @@ namespace C_R_M.Controllers
         // GET: Empresas/Create
         public ActionResult Create()
         {
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
             ViewBag.Id_Canton = new SelectList(db.Canton.Where(a => a.Id_Canton == 1), "Id_Canton", "Nombre");
             ViewBag.Id_Distrito = new SelectList(db.Distrito.Where(a => a.Id_Distrito == 1), "Id_Distrito", "Nombre");
             ViewBag.Id_Provincia = new SelectList(db.Provincia.Where(a => a.Id_Provincia == 1), "Id_Provincia", "Nombre");
             ViewBag.Pais = new SelectList(db.Pais, "Id_Pais", "Nombre");
-            return View();
+            Empresa empresa = new Empresa { Id_Pais = db.Pais.First(p => p.Nombre.Equals("Costa Rica")).Id_Pais };
+            return View(empresa);
         }
 
         // POST: Empresas/Create
@@ -51,34 +61,71 @@ namespace C_R_M.Controllers
         // más información vea https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "Id_Empresa,Nombre,Correo,Tipo_Cedula,Cedula,Pais,Id_Provincia,Id_Canton,Id_Distrito,Otras_Señas,Codigo_Postal")] Empresa empresa)
+        public ActionResult Create([Bind(Include = "Id_Empresa,Nombre,Correo,Tipo_Cedula,Cedula,Id_Pais,Id_Provincia,Id_Canton,Id_Distrito,Otras_Señas,Codigo_Postal")] Empresa empresa)
         {
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
             try
             {
                 if (ModelState.IsValid)
                 {
+                    if (empresa.Id_Pais == 48 && empresa.Id_Provincia == 0)
+                    {
+                        ModelState.AddModelError("Id_Provincia", "Seleccione una Provincia");
+
+                        ViewBag.Id_Canton = new SelectList(db.Canton, "Id_Canton", "Nombre", empresa.Id_Canton);
+                        ViewBag.Id_Distrito = new SelectList(db.Distrito, "Id_Distrito", "Nombre", empresa.Id_Distrito);
+                        ViewBag.Id_Provincia = new SelectList(db.Provincia, "Id_Provincia", "Nombre", empresa.Id_Provincia);
+                        ViewBag.Pais = new SelectList(db.Pais, "Id_Pais", "Nombre", empresa.Id_Pais);
+                        return View(empresa);
+                    }
+
+                    if (empresa.Id_Pais == 48 && empresa.Id_Canton == 0)
+                    {
+                        ModelState.AddModelError("Id_Canton", "Seleccione un Cantón");
+
+                        ViewBag.Id_Canton = new SelectList(db.Canton, "Id_Canton", "Nombre", empresa.Id_Canton);
+                        ViewBag.Id_Distrito = new SelectList(db.Distrito, "Id_Distrito", "Nombre", empresa.Id_Distrito);
+                        ViewBag.Id_Provincia = new SelectList(db.Provincia, "Id_Provincia", "Nombre", empresa.Id_Provincia);
+                        ViewBag.Pais = new SelectList(db.Pais, "Id_Pais", "Nombre", empresa.Id_Pais);
+                        return View(empresa);
+                    }
+
+                    if (empresa.Id_Pais == 48 && empresa.Id_Distrito ==0 )
+                    {
+                        ModelState.AddModelError("Id_Distrito", "Seleccione un Distrito");
+
+                        ViewBag.Id_Canton = new SelectList(db.Canton, "Id_Canton", "Nombre", empresa.Id_Canton);
+                        ViewBag.Id_Distrito = new SelectList(db.Distrito, "Id_Distrito", "Nombre", empresa.Id_Distrito);
+                        ViewBag.Id_Provincia = new SelectList(db.Provincia, "Id_Provincia", "Nombre", empresa.Id_Provincia);
+                        ViewBag.Pais = new SelectList(db.Pais, "Id_Pais", "Nombre", empresa.Id_Pais);
+                        return View(empresa);
+                    }
+                    
+                    
                     db.Empresa.Add(empresa);
                     db.SaveChanges();
+                    
                     return RedirectToAction("Index");
                 }
-
-                ViewBag.Id_Canton = new SelectList(db.Canton, "Id_Canton", "Nombre", empresa.Id_Canton);
-                ViewBag.Id_Distrito = new SelectList(db.Distrito, "Id_Distrito", "Nombre", empresa.Id_Distrito);
-                ViewBag.Id_Provincia = new SelectList(db.Provincia, "Id_Provincia", "Nombre", empresa.Id_Provincia);
-                ViewBag.Pais = new SelectList(db.Pais, "Id_Pais", "Nombre", empresa.Pais);
-                return View(empresa);
             }
-            catch (Exception)
+            catch (SqlException)
             {
-
-                throw;
+                return RedirectToAction("Index");
             }
-            
+
+            ViewBag.Id_Canton = new SelectList(db.Canton, "Id_Canton", "Nombre", empresa.Id_Canton);
+            ViewBag.Id_Distrito = new SelectList(db.Distrito, "Id_Distrito", "Nombre", empresa.Id_Distrito);
+            ViewBag.Id_Provincia = new SelectList(db.Provincia, "Id_Provincia", "Nombre", empresa.Id_Provincia);
+            ViewBag.Pais = new SelectList(db.Pais, "Id_Pais", "Nombre", empresa.Id_Pais);
+            return View(empresa);
         }
 
         // GET: Empresas/Edit/5
         public ActionResult Edit(int? id)
         {
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -91,7 +138,7 @@ namespace C_R_M.Controllers
             ViewBag.Id_Canton = new SelectList(db.Canton.Where(c => c.Provincia == empresa.Id_Provincia && c.Id_Canton != 1).OrderBy(a => a.Nombre), "Id_Canton", "Nombre");
             ViewBag.Id_Distrito = new SelectList(db.Distrito.Where(c => c.Canton == empresa.Id_Canton && c.Id_Distrito != 1).OrderBy(a => a.Nombre), "Id_Distrito", "Nombre");
             ViewBag.Id_Provincia = new SelectList(db.Provincia.Where(a => a.Id_Provincia != 1).OrderBy(a => a.Nombre), "Id_Provincia", "Nombre");
-            ViewBag.Pais = new SelectList(db.Pais, "Id_Pais", "Nombre", empresa.Pais);
+            ViewBag.Pais = new SelectList(db.Pais, "Id_Pais", "Nombre", empresa.Id_Pais);
             return View(empresa);
         }
 
@@ -100,10 +147,45 @@ namespace C_R_M.Controllers
         // más información vea https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "Id_Empresa,Nombre,Correo,Tipo_Cedula,Cedula,Pais,Id_Provincia,Id_Canton,Id_Distrito,Otras_Señas,Codigo_Postal")] Empresa empresa)
+        public ActionResult Edit([Bind(Include = "Id_Empresa,Nombre,Correo,Tipo_Cedula,Cedula,Id_Pais,Id_Provincia,Id_Canton,Id_Distrito,Otras_Señas,Codigo_Postal")] Empresa empresa)
         {
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
             if (ModelState.IsValid)
             {
+                if (empresa.Id_Pais == 48 && empresa.Id_Provincia == 0)
+                {
+                    ModelState.AddModelError("Id_Provincia", "Seleccione una Provincia");
+
+                    ViewBag.Id_Canton = new SelectList(db.Canton, "Id_Canton", "Nombre", empresa.Id_Canton);
+                    ViewBag.Id_Distrito = new SelectList(db.Distrito, "Id_Distrito", "Nombre", empresa.Id_Distrito);
+                    ViewBag.Id_Provincia = new SelectList(db.Provincia, "Id_Provincia", "Nombre", empresa.Id_Provincia);
+                    ViewBag.Pais = new SelectList(db.Pais, "Id_Pais", "Nombre", empresa.Id_Pais);
+                    return View(empresa);
+                }
+
+                if (empresa.Id_Pais == 48 && empresa.Id_Canton == 0)
+                {
+                    ModelState.AddModelError("Id_Canton", "Seleccione un Cantón");
+
+                    ViewBag.Id_Canton = new SelectList(db.Canton, "Id_Canton", "Nombre", empresa.Id_Canton);
+                    ViewBag.Id_Distrito = new SelectList(db.Distrito, "Id_Distrito", "Nombre", empresa.Id_Distrito);
+                    ViewBag.Id_Provincia = new SelectList(db.Provincia, "Id_Provincia", "Nombre", empresa.Id_Provincia);
+                    ViewBag.Pais = new SelectList(db.Pais, "Id_Pais", "Nombre", empresa.Id_Pais);
+                    return View(empresa);
+                }
+
+                if (empresa.Id_Pais == 48 && empresa.Id_Distrito == 0)
+                {
+                    ModelState.AddModelError("Id_Distrito", "Seleccione un Distrito");
+
+                    ViewBag.Id_Canton = new SelectList(db.Canton, "Id_Canton", "Nombre", empresa.Id_Canton);
+                    ViewBag.Id_Distrito = new SelectList(db.Distrito, "Id_Distrito", "Nombre", empresa.Id_Distrito);
+                    ViewBag.Id_Provincia = new SelectList(db.Provincia, "Id_Provincia", "Nombre", empresa.Id_Provincia);
+                    ViewBag.Pais = new SelectList(db.Pais, "Id_Pais", "Nombre", empresa.Id_Pais);
+                    return View(empresa);
+                }
+
                 db.Entry(empresa).State = EntityState.Modified;
                 db.SaveChanges();
                 return RedirectToAction("Index");
@@ -111,13 +193,15 @@ namespace C_R_M.Controllers
             ViewBag.Id_Canton = new SelectList(db.Canton, "Id_Canton", "Nombre", empresa.Id_Canton);
             ViewBag.Id_Distrito = new SelectList(db.Distrito, "Id_Distrito", "Nombre", empresa.Id_Distrito);
             ViewBag.Id_Provincia = new SelectList(db.Provincia, "Id_Provincia", "Nombre", empresa.Id_Provincia);
-            ViewBag.Pais = new SelectList(db.Pais, "Id_Pais", "Nombre", empresa.Pais);
+            ViewBag.Pais = new SelectList(db.Pais, "Id_Pais", "Nombre", empresa.Id_Pais);
             return View(empresa);
         }
 
         // GET: Empresas/Delete/5
         public ActionResult Delete(int? id)
         {
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -135,6 +219,8 @@ namespace C_R_M.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
             Empresa empresa = db.Empresa.Find(id);
             db.Empresa.Remove(empresa);
             db.SaveChanges();
@@ -150,15 +236,16 @@ namespace C_R_M.Controllers
             base.Dispose(disposing);
         }
 
-
         [HttpPost]
+
         public JsonResult ListCantones(Int64 ID_PROVINCIA, CRMEntities _Model)
         {
-            ViewBag.e = new SelectList(db.Canton.Where(c => c.Provincia == ID_PROVINCIA && c.Id_Canton != 1 ).OrderBy(a => a.Nombre), "Id_Canton", "Nombre");
+            ViewBag.e = new SelectList(db.Canton.Where(c => c.Provincia == ID_PROVINCIA && c.Id_Canton != 1).OrderBy(a => a.Nombre), "Id_Canton", "Nombre");
             return Json(ViewBag.e);
         }
 
         [HttpPost]
+
         public JsonResult ListDistritos(Int64 ID_CANTON, CRMEntities _Model)
         {
             ViewBag.e = new SelectList(db.Distrito.Where(c => c.Canton == ID_CANTON && c.Id_Distrito != 1).OrderBy(a => a.Nombre), "Id_Distrito", "Nombre");
@@ -166,15 +253,15 @@ namespace C_R_M.Controllers
         }
 
         [HttpPost]
-        public JsonResult ListProvincias(Int64 ID_PAIS,  CRMEntities _Model)
+        public JsonResult ListProvincias(Int64 ID_PAIS, CRMEntities _Model)
         {
-            ViewBag.e = new SelectList(db.Provincia.Where(a => a.Id_Provincia == 1), "Id_Provincia", "Nombre");
+            ViewBag.e = new SelectList(db.Provincia.Where(a => a.Id_Provincia == 0), "Id_Provincia", "Nombre");
             if (ID_PAIS == 48)
             {
                 ViewBag.e = new SelectList(db.Provincia.Where(a => a.Id_Provincia != 1).OrderBy(a => a.Nombre), "Id_Provincia", "Nombre");
             }
-            
             return Json(ViewBag.e);
+
         }
     }
 }
